working code extracted
